# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: واجهة طلبات بيانات إعلانات الوفاة
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: requests
  - name: admin
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /obituary-requests:
    get:
      operationId: listObituaryRequests
      tags: [requests]
      description: للمسؤول فقط
      responses:
        "200":
          description: قائمة الطلبات من الأحدث
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ObituaryRequest"
        "401":
          description: يلزم تسجيل الدخول كمسؤول
    post:
      operationId: createObituaryRequest
      tags: [requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObituaryRequestInput"
      responses:
        "201":
          description: تم إنشاء الطلب
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObituaryRequest"
  /obituary-requests/{requestNumber}:
    get:
      operationId: getObituaryRequest
      tags: [requests]
      parameters:
        - $ref: "#/components/parameters/RequestNumber"
      responses:
        "200":
          description: تفاصيل الطلب
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObituaryRequest"
        "404":
          description: الطلب غير موجود
        "429":
          description: محاولات بحث كثيرة، حاول لاحقاً
    put:
      operationId: updateObituaryRequest
      tags: [requests]
      parameters:
        - $ref: "#/components/parameters/RequestNumber"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObituaryRequestUpdate"
      description: للمسؤول فقط
      responses:
        "200":
          description: تم تحديث الطلب
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObituaryRequest"
        "401":
          description: يلزم تسجيل الدخول كمسؤول
        "404":
          description: الطلب غير موجود
    delete:
      operationId: deleteObituaryRequest
      tags: [requests]
      description: للمسؤول فقط
      parameters:
        - $ref: "#/components/parameters/RequestNumber"
      responses:
        "204":
          description: تم حذف الطلب
        "401":
          description: يلزم تسجيل الدخول كمسؤول
        "404":
          description: الطلب غير موجود
  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      responses:
        "200":
          description: حالة جلسة المسؤول
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: تم تسجيل الدخول (تُرسل الجلسة في ملف تعريف ارتباط HttpOnly)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: كلمة المرور غير صحيحة
        "429":
          description: محاولات خاطئة كثيرة، حاول لاحقاً
        "503":
          description: كلمة مرور المسؤول لم تُضبط في الخادم أو ضعيفة
  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      responses:
        "200":
          description: تم تسجيل الخروج
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
components:
  parameters:
    RequestNumber:
      name: requestNumber
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    AdminSession:
      type: object
      required: [authenticated, configured]
      properties:
        authenticated:
          type: boolean
        configured:
          type: boolean
          description: كلمة مرور المسؤول مضبوطة في الخادم وقوية بما يكفي
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          maxLength: 200
    LinkedPerson:
      type: object
      description: شخص مرجعي (الزوج، الأب، الأب المشترك، أو مرجع «أبناء /») مع حالته
      properties:
        title:
          type: string
        name:
          type: string
        deceased:
          type: boolean
    SpouseReference:
      type: object
      description: زوج المتوفاة؛ النوع يحدد «حرم» أو «أرملة»
      properties:
        kind:
          type: string
          enum: [harem, widow]
        title:
          type: string
        name:
          type: string
        deceased:
          type: boolean
    DeceasedPerson:
      type: object
      required: [gender]
      properties:
        fullName:
          type: string
        gender:
          type: string
          enum: [man, woman, boy, girl, other]
        identifyBy:
          type: string
          description: طريقة التعريف بالمتوفى في رأس الإعلان
          enum: [name, kunya, spouse, father, children]
        kunya:
          type: string
        age:
          type: ["integer", "null"]
        ageUnit:
          type: string
          enum: [years, months, days]
        nationality:
          type: string
        deathPlace:
          type: string
        title:
          type: string
        occupation:
          type: string
        note:
          type: string
        noChildren:
          type: boolean
        spouse:
          $ref: "#/components/schemas/SpouseReference"
        father:
          $ref: "#/components/schemas/LinkedPerson"
    RelativePerson:
      type: object
      required: [name, deceased]
      properties:
        name:
          type: string
        occupation:
          type: string
        deceased:
          type: boolean
    RelativeGroup:
      type: object
      required: [relation, people]
      properties:
        relation:
          type: string
        relationKey:
          type: string
          description: صلة الأشخاص بالمتوفى؛ يُشتق منها عنوان الإعلان مثل «والدة كل من»
          enum:
            - children
            - full_siblings
            - siblings
            - grandchildren
            - brothers_children
            - sisters_children
            - father
            - grandfather
            - paternal_uncles
            - maternal_uncles
            - daughters_husbands
            - sisters_husbands
            - other
        familyReference:
          type: string
        reference:
          $ref: "#/components/schemas/LinkedPerson"
        deceasedPlacement:
          type: string
          enum: [auto, inline, grouped]
        deceasedIndex:
          type: ["integer", "null"]
        people:
          type: array
          items:
            $ref: "#/components/schemas/RelativePerson"
    PrayerDetails:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        day:
          type: string
        weekday:
          type: string
        time:
          type: string
        place:
          type: string
        mapLink:
          type: string
    BurialDetails:
      type: object
      required: [status, outsideQatar]
      properties:
        status:
          type: string
          enum: [upcoming, completed, postponed]
        day:
          type: string
        weekday:
          type: string
        time:
          type: string
        cemetery:
          type: string
        mapLink:
          type: string
        outsideQatar:
          type: boolean
        outsideLocation:
          type: string
        note:
          type: string
          description: ملاحظة الدفن (سبب التأجيل، أو وصف دفن تمّ مثل «تم الدفن في مكة المكرمة»)
    CondolenceScheduleEntry:
      type: object
      properties:
        days:
          type: string
        time:
          type: string
    CondolenceCard:
      type: object
      required: [audience]
      properties:
        audience:
          type: string
          enum: [men, women]
        location:
          type: string
        mapLink:
          type: string
        start:
          type: string
        durationDays:
          type: ["integer", "null"]
        time:
          type: string
        houseNumber:
          type: string
        buildingNumber:
          type: string
        street:
          type: string
        area:
          type: string
        floor:
          type: string
        apartmentNumber:
          type: string
        locationNotes:
          type: string
        until:
          type: string
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/CondolenceScheduleEntry"
        deceasedIndex:
          type: ["integer", "null"]
    CondolencePhoneContact:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
    CondolenceCancellation:
      type: object
      properties:
        audience:
          type: string
          enum: [men, women, all]
        from:
          type: string
        reason:
          type: string
        phoneOnly:
          type: boolean
    ObituaryRequestInput:
      type: object
      required:
        - deceasedPeople
        - relatives
        - prayer
        - burial
        - condolenceOptions
        - condolences
        - condolencePhoneContacts
      properties:
        messageType:
          type: string
          description: نوع الرسالة (إعلان، تأجيل، تعديل، إلغاء عزاء)
          enum: [announcement, postponement, amendment, condolence_cancellation]
        relatedRequestNumber:
          type: string
        announcementMode:
          type: string
          description: صيغة الإعلان عند تعدد المتوفين
          enum: [single, unrelated, siblings, father_first, mother_child]
        sharedParent:
          $ref: "#/components/schemas/LinkedPerson"
        cancellation:
          $ref: "#/components/schemas/CondolenceCancellation"
        deceasedPeople:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/DeceasedPerson"
        relatives:
          type: array
          items:
            $ref: "#/components/schemas/RelativeGroup"
        noRelatives:
          type: boolean
          description: أكّد المرسل أن المتوفى ليس لديه أقارب يُذكرون
        prayer:
          $ref: "#/components/schemas/PrayerDetails"
        burial:
          $ref: "#/components/schemas/BurialDetails"
        condolenceOptions:
          type: array
          items:
            type: string
            enum: [phone, men, women, men_cemetery, tbd]
        phoneAudience:
          type: string
          enum: [all, men, women]
        condolenceNote:
          type: string
        condolences:
          type: array
          items:
            $ref: "#/components/schemas/CondolenceCard"
        condolencePhoneContacts:
          type: array
          items:
            $ref: "#/components/schemas/CondolencePhoneContact"
        notes:
          type: string
    ObituaryRequestUpdate:
      allOf:
        - $ref: "#/components/schemas/ObituaryRequestInput"
        - type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [new, reviewing, ready, completed]
    ObituaryRequest:
      allOf:
        - $ref: "#/components/schemas/ObituaryRequestInput"
        - type: object
          required: [id, requestNumber, createdAt, updatedAt, status]
          properties:
            id:
              type: integer
            requestNumber:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            status:
              type: string
              enum: [new, reviewing, ready, completed]