# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * واجهة طلبات بيانات إعلانات الوفاة
 * OpenAPI spec version: 0.3.0
 */

export interface AdminLoginInput {
  /** @maxLength 200 */
  password: string;
}