# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * واجهة طلبات بيانات إعلانات الوفاة
 * OpenAPI spec version: 0.3.0
 */

export interface AdminSession {
  authenticated: boolean;
  /** كلمة مرور المسؤول مضبوطة في الخادم وقوية بما يكفي */
  configured: boolean;
}